Website includes our Privacy Policy
From Sharon:

> I need to include the Privacy Policy in Spark [our recruitment
vendor], and it also needs to be part of our website.

This commit introduces our privacy policy. We don't link to it from
anywhere on our website: we publish it as a standalone page.

The privacy policy was originally a Gdoc which I downloaded as HTML and
then compared against the original, adding a few minor stylistic fixes.
Note that this HTML is horribly bloated, no doubt from the Gdoc
conversion, and it was all I could do to restrain myself from re-doing
the HTML from scratch.

Signed-off-by: Brian Cunnie <cunnie@majestic-labs.ai>